Opt component deleted
